Support image attachments when creating recurring items
When deleting items we don't delete image attachments anyways, so no concerns there, and when marking for upload we only do that once for the image, so no concerns there. Only concern is when editing an item, if they remove the image via editing, we will delete and that'll delete the shared reference, but oh well, niche case. Someday should improve image logic.